Bring pcornet v3 field attributes into compliance
Mostly varchar and numeric fields. Note that uniformly, if fields are
specified in the pcornet doc as "NUMBER(8)", they are here given
"number,,10,2", to allow for some decimal specification and avoid the
application of a default scale which would occur if the scale were
left off or specified as 0.

Fixes #136

Signed-off-by: Aaron Browne <aaron0browne@gmail.com>