Make circle.yml fork-friendly
The main change is the introduction of USER_NAME
(CIRCLE_PROJECT_USERNAME) to avoid using the original repo's github
organization in URLs.

Other minor changes:

* dms runs with more verbose logging.
* Errors when starting dms or dmsa are caught directly.
* Gross failures of dms and dmsa containers are caught directly.
* If dms or dmsa do not respond, the container logs are dumped.
* dms and dmsa logs are saved as artifacts.

For PRs of forks to be tested, CircleCI advanced settings need to be
set:

* "Build forked pull requests" => "On"
* "Pass secrets to builds from forked pull requests" => "Off"

The latter may have no effect right now, but it seems needed for
safety.